+++ src/features/memo/components/MemoList/MemoListHeader.tsx
@@ -1,6 +1,7 @@
-import { ChevronLeft, ChevronRight, Trash2 } from 'lucide-react';
+import { Trash2 } from 'lucide-react';
 import { Button } from '@/shared/components/shadcn/button';
 import type { Memo } from '@/features/memo/types/memoTypes';
+import PaginationNav from '@/shared/components/ui/PaginationNav';
 
 interface MemoListHeaderProps {
   memos?: Memo[];
@@ -22,15 +23,6 @@ const MemoListHeader = ({
   const selectedCount = selectedRows.size;
   const totalCount = memos?.length ?? 0;
 
-  const isFirstPage = currentPage === 0;
-  const isLastPage = currentPage + 1 >= pageCount;
-
-  const goPrevPage = () => setCurrentPage((p) => p - 1);
-  const goNextPage = () => setCurrentPage((p) => p + 1);
-
-  const paginationBtnClass =
-    'rounded-full border disabled:border-gray-500 disabled:bg-white disabled:text-gray-600';
-
   return (
     <header className="flex items-center justify-between px-2">
       <div className="flex items-center gap-3 label2-regular md:label1-regular text-gray-500">
@@ -51,27 +43,11 @@ const MemoListHeader = ({
           <span>총 {totalCount}개 메모</span>
         )}
       </div>
-
-      <nav className="flex items-center gap-2 pr-2">
-        <Button
-          variant="defaultBoost"
-          size="icon"
-          disabled={isFirstPage}
-          onClick={goPrevPage}
-          className={paginationBtnClass}
-        >
-          <ChevronLeft />
-        </Button>
-        <Button
-          variant="defaultBoost"
-          size="icon"
-          disabled={isLastPage}
-          onClick={goNextPage}
-          className={paginationBtnClass}
-        >
-          <ChevronRight />
-        </Button>
-      </nav>
+      <PaginationNav
+        currentPage={currentPage}
+        pageCount={pageCount}
+        setCurrentPage={setCurrentPage}
+      />
     </header>
   );
 };
